requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.2
certifi==2020.11.8
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyTelegramBotAPI==3.7.4
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
six==1.15.0
telepot==12.7
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.6
yarl==1.6.3

telegram~=0.0.1
python-dotenv~=0.17.1